Cleaned up and simplified code greatly
Removed the redundant Maybe's from getFirstNum and getClosest.

Moved getClosest into getFirstNum since it only really makes sense in
the context of that function.

Switched to using a match statement for getClosest instead of an if/else.

Removed strToNum in favor of numberMap. This makes things a lot easier
since we already know what the associated Char should be for any given
number we want to search with.

Switched names of variables to more descriptive ones.